+++ packages/ubs/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.1.13]
+### Fixed
+- UBReport: in case `$fn` function argument is empty return empty string instead of `null`
+
 ## [5.1.12]
 ### Fixed
 - UBReport: HTML report will replace `<!-- pagebreak -->` placeholder to special element before print